refactor(tui): extract tutorial state boundary
